Modify staticPage.py file for MVC
